pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "umi-dex"
version = "0.2.0"
description = "Offline post-processing pipeline for ROS-recorded UMI-Dex bags: ORB-SLAM3 replay, CAN decoding, video encoding, and aligned dataset assembly."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24",
    "opencv-python>=4.8",
    "orbslam3-python>=2.0.0",
    "rosbags>=0.10",
    "av>=12.0",
    "pandas>=2.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
umi-inspect = "umi_dex.cli.inspect:main"
umi-extract = "umi_dex.cli.extract:main"
umi-slam = "umi_dex.cli.slam:main"
umi-process = "umi_dex.cli.process:main"

[tool.hatch.build.targets.wheel]
packages = ["src/umi_dex"]
